refactor(NavItem): prefer dedicated folder for list items
