+++ website-frontend/src/lib/components/list_items/NavItem.svelte
